ai-rep-coach/requirements.txt
streamlit
mediapipe
opencv-python
numpy
google-genai
python-dotenv
